Improve KeyManager Implementation
