+++ components/apimgt/org.wso2.carbon.apimgt.core/src/main/java/org/wso2/carbon/apimgt/core/api/KeyManager.java
@@ -44,7 +44,7 @@ public interface KeyManager {
      *
      * @param oauthAppRequest - this object contains values of oAuth app properties.
      * @return OAuthApplicationInfo object with oAuthApplication properties.
-     * @throws KeyManagementException   Error while creating application.
+     * @throws KeyManagementException Error while creating application.
      */
     OAuthApplicationInfo createApplication(OAuthAppRequest oauthAppRequest) throws KeyManagementException;
 
@@ -53,7 +53,7 @@ public interface KeyManager {
      *
      * @param oAuthApplicationInfo Up-to-date information of OAuth Application
      * @return OAuthApplicationInfo this object will  contain all the properties of updated oAuth application
-     * @throws KeyManagementException   Error while updating application.
+     * @throws KeyManagementException Error while updating application.
      */
     OAuthApplicationInfo updateApplication(OAuthApplicationInfo oAuthApplicationInfo) throws KeyManagementException;
 
@@ -61,7 +61,7 @@ public interface KeyManager {
      * Delete auth application
      *
      * @param consumerKey - will take consumer key as parameter
-     * @throws KeyManagementException   Error while deleting application.
+     * @throws KeyManagementException Error while deleting application.
      */
     void deleteApplication(String consumerKey) throws KeyManagementException;
 
@@ -70,7 +70,7 @@ public interface KeyManager {
      *
      * @param consumerKey will take consumer key as parameter
      * @return json string
-     * @throws KeyManagementException   Error while retrieving application.
+     * @throws KeyManagementException Error while retrieving application.
      */
     OAuthApplicationInfo retrieveApplication(String consumerKey) throws KeyManagementException;
 
@@ -87,9 +87,10 @@ public interface KeyManager {
     /**
      * Get details about an access token. As a part of the response, consumer key against which token was obtained
      * must be returned.
-     * @param accessToken   Access token which needs to be check.
+     *
+     * @param accessToken Access token which needs to be check.
      * @return {@code AccessTokenInfo}
-     * @throws KeyManagementException   Error while retrieving token meta data.
+     * @throws KeyManagementException Error while retrieving token meta data.
      */
     AccessTokenInfo getTokenMetaData(String accessToken) throws KeyManagementException;
 
@@ -113,8 +114,9 @@ public interface KeyManager {
 
     /**
      * Load the key manager configuration
-     * @param configuration   configuration object.
-     * @throws KeyManagementException   Error while loading configs.
+     *
+     * @param configuration configuration object.
+     * @throws KeyManagementException Error while loading configs.
      */
     void loadConfiguration(KeyManagerConfiguration configuration) throws KeyManagementException;
 
@@ -124,7 +126,7 @@ public interface KeyManager {
      * @param api                this is a API object which contains all details about a API.
      * @param resourceAttributes this param will contains additional details if required.
      * @return true if sucessfully registered. false if there is a error while registering a new resource.
-     * @throws KeyManagementException   Error while registering new resource.
+     * @throws KeyManagementException Error while registering new resource.
      */
     boolean registerNewResource(API api, Map resourceAttributes) throws KeyManagementException;
 
@@ -133,7 +135,7 @@ public interface KeyManager {
      *
      * @param apiId APIM api id.
      * @return It will return a Map with registered resource details.
-     * @throws KeyManagementException   Error while retrieving resource by ID.
+     * @throws KeyManagementException Error while retrieving resource by ID.
      */
     Map getResourceByApiId(String apiId) throws KeyManagementException;
 
@@ -143,7 +145,7 @@ public interface KeyManager {
      * @param api                this is a API object which contains all details about a API.
      * @param resourceAttributes this param will contains additional details if required.
      * @return TRUE|FALSE. if it is successfully updated it will return TRUE or else FALSE.
-     * @throws KeyManagementException   Error while updating resource.
+     * @throws KeyManagementException Error while updating resource.
      */
     boolean updateRegisteredResource(API api, Map resourceAttributes) throws KeyManagementException;
 
@@ -151,15 +153,15 @@ public interface KeyManager {
      * This method will accept API id  as a parameter  and will delete the registered resource.
      *
      * @param apiID API id.
-     * @throws KeyManagementException   Error while deleting resource.
+     * @throws KeyManagementException Error while deleting resource.
      */
     void deleteRegisteredResourceByAPIId(String apiID) throws KeyManagementException;
 
     /**
      * This method will be used to delete mapping records of oAuth applications.
      *
-     * @param consumerKey   Key of the application
-     * @throws KeyManagementException  Error while deleting mapped application.
+     * @param consumerKey Key of the application
+     * @throws KeyManagementException Error while deleting mapped application.
      */
     void deleteMappedApplication(String consumerKey) throws KeyManagementException;
