+++ components/apimgt/org.wso2.carbon.apimgt.core/src/main/java/org/wso2/carbon/apimgt/core/auth/dto/ScopeInfo.java
@@ -42,7 +42,6 @@ public final class ScopeInfo {
 
     public void setName(String name) {
         this.name = name;
-        this.displayName = name;
     }
 
     public String getDescription() {
