+++ components/apimgt/org.wso2.carbon.apimgt.core/src/test/java/org/wso2/carbon/apimgt/core/impl/DefaultIdentityProviderImplTestCase.java
@@ -24,10 +24,7 @@ import org.mockito.Mockito;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.wso2.carbon.apimgt.core.api.UserNameMapper;
-import org.wso2.carbon.apimgt.core.auth.DCRMServiceStub;
-import org.wso2.carbon.apimgt.core.auth.OAuth2ServiceStubs;
 import org.wso2.carbon.apimgt.core.auth.SCIMServiceStub;
-import org.wso2.carbon.apimgt.core.auth.ScopeRegistrationServiceStub;
 import org.wso2.carbon.apimgt.core.auth.dto.SCIMUser;
 import org.wso2.carbon.apimgt.core.exception.IdentityProviderException;
 import org.wso2.carbon.apimgt.core.models.User;
@@ -44,16 +41,12 @@ public class DefaultIdentityProviderImplTestCase {
     @Test
     public void testGetIdOfUser() throws Exception {
         SCIMServiceStub scimServiceStub = Mockito.mock(SCIMServiceStub.class);
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         UserNameMapper userNameMapper = Mockito.mock(UserNameMapperImpl.class);
         Mockito.when(userNameMapper.getLoggedInUserIDFromPseudoName("invalid_user")).thenReturn("invalid_user");
         Mockito.when(userNameMapper.getLoggedInUserIDFromPseudoName("John")).thenReturn("John");
         Mockito.when(userNameMapper.getLoggedInUserIDFromPseudoName("invalid_user_giving_null_response")).
                 thenReturn("invalid_user_giving_null_response");
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, dcrmServiceStub,
-                oAuth2ServiceStub, scopeRegistrationServiceStub, userNameMapper);
+        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, userNameMapper);
 
         String validUserName = "John";
         final String validUserSearchQuery = "userName Eq " + validUserName;
@@ -113,12 +106,8 @@ public class DefaultIdentityProviderImplTestCase {
     @Test
     public void testGetRoleNamesOfUser() throws Exception {
         SCIMServiceStub scimServiceStub = Mockito.mock(SCIMServiceStub.class);
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
         UserNameMapper userNameMapper = Mockito.mock(UserNameMapperImpl.class);
-        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, dcrmServiceStub,
-                oAuth2ServiceStub, scopeRegistrationServiceStub, userNameMapper);
+        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, userNameMapper);
 
         String validUserId = "a42b4760-120d-432e-8042-4a7f12e3346c";
         String roleName1 = "subscriber";
@@ -195,12 +184,8 @@ public class DefaultIdentityProviderImplTestCase {
     @Test
     public void testIsValidRole() throws Exception {
         SCIMServiceStub scimServiceStub = Mockito.mock(SCIMServiceStub.class);
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         UserNameMapper userNameMapper = Mockito.mock(UserNameMapperImpl.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, dcrmServiceStub,
-                oAuth2ServiceStub, scopeRegistrationServiceStub, userNameMapper);
+        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, userNameMapper);
 
         final String validRole = "engineer";
         final String validRoleSearchQuery = "displayName Eq " + validRole;
@@ -222,12 +207,8 @@ public class DefaultIdentityProviderImplTestCase {
     @Test
     public void testGetRoleIdsOfUser() throws Exception {
         SCIMServiceStub scimServiceStub = Mockito.mock(SCIMServiceStub.class);
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         UserNameMapper userNameMapper = Mockito.mock(UserNameMapperImpl.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, dcrmServiceStub,
-                oAuth2ServiceStub, scopeRegistrationServiceStub, userNameMapper);
+        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, userNameMapper);
 
         String validUserId = "a42b4760-120d-432e-8042-4a7f12e3346c";
         String roleName1 = "subscriber";
@@ -304,12 +285,8 @@ public class DefaultIdentityProviderImplTestCase {
     @Test
     public void testGetRoleId() throws Exception {
         SCIMServiceStub scimServiceStub = Mockito.mock(SCIMServiceStub.class);
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         UserNameMapper userNameMapper = Mockito.mock(UserNameMapperImpl.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, dcrmServiceStub,
-                oAuth2ServiceStub, scopeRegistrationServiceStub, userNameMapper);
+        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, userNameMapper);
 
         String validRoleName = "engineer";
         final String validRoleSearchQuery = "displayName Eq " + validRoleName;
@@ -369,13 +346,9 @@ public class DefaultIdentityProviderImplTestCase {
     @Test
     public void testGetRoleName() throws Exception {
         SCIMServiceStub scimServiceStub = Mockito.mock(SCIMServiceStub.class);
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         UserNameMapper userNameMapper = Mockito.mock(UserNameMapperImpl.class);
 
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, dcrmServiceStub,
-                oAuth2ServiceStub, scopeRegistrationServiceStub, userNameMapper);
+        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, userNameMapper);
 
         String validRoleId = "ac093278-9343-466c-8a71-af47921a575b";
         String expectedRoleName = "engineer";
@@ -443,13 +416,8 @@ public class DefaultIdentityProviderImplTestCase {
     @Test
     public void testRegisterUser() throws Exception {
         SCIMServiceStub scimServiceStub = Mockito.mock(SCIMServiceStub.class);
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         UserNameMapper userNameMapper = Mockito.mock(UserNameMapperImpl.class);
-
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, dcrmServiceStub,
-                oAuth2ServiceStub, scopeRegistrationServiceStub, userNameMapper);
+        DefaultIdentityProviderImpl idpImpl = new DefaultIdentityProviderImpl(scimServiceStub, userNameMapper);
 
         //happy path
         User user = new User();
