+++ components/apimgt/org.wso2.carbon.apimgt.core/src/test/java/org/wso2/carbon/apimgt/core/impl/DefaultKeyManagerImplTestCase.java
@@ -27,28 +27,24 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.wso2.carbon.apimgt.core.auth.DCRMServiceStub;
 import org.wso2.carbon.apimgt.core.auth.OAuth2ServiceStubs;
-import org.wso2.carbon.apimgt.core.auth.ScopeRegistrationServiceStub;
+import org.wso2.carbon.apimgt.core.auth.ScopeRegistration;
 import org.wso2.carbon.apimgt.core.auth.dto.DCRClientInfo;
 import org.wso2.carbon.apimgt.core.auth.dto.OAuth2IntrospectionResponse;
 import org.wso2.carbon.apimgt.core.auth.dto.OAuth2TokenInfo;
-import org.wso2.carbon.apimgt.core.auth.dto.ScopeInfo;
-import org.wso2.carbon.apimgt.core.exception.APIManagementException;
 import org.wso2.carbon.apimgt.core.exception.KeyManagementException;
 import org.wso2.carbon.apimgt.core.models.AccessTokenInfo;
 import org.wso2.carbon.apimgt.core.models.AccessTokenRequest;
 import org.wso2.carbon.apimgt.core.models.OAuthAppRequest;
 import org.wso2.carbon.apimgt.core.models.OAuthApplicationInfo;
-import org.wso2.carbon.apimgt.core.models.Scope;
 import org.wso2.carbon.apimgt.core.util.KeyManagerConstants;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 
 import static org.mockito.Matchers.any;
 
-public class WSO2ISKeyManagerImplTestCase {
+public class DefaultKeyManagerImplTestCase {
 
     private static final String consumerKey = "xxx-xxx-xxx-xxx";
     private static final String consumerSecret = "yyy-yyy-yyy-yyy";
@@ -57,9 +53,8 @@ public class WSO2ISKeyManagerImplTestCase {
     public void testCreateApplication() throws Exception {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock(ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub, scopeRegistration);
 
         //happy path - 201
         ////request object to key manager
@@ -149,9 +144,10 @@ public class WSO2ISKeyManagerImplTestCase {
     public void testUpdateApplication() throws Exception {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
         final String consumerKey = "xxx-xxx-xxx-xxx";
 
         //happy path - 200
@@ -244,9 +240,10 @@ public class WSO2ISKeyManagerImplTestCase {
     public void testDeleteApplication() throws Exception {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
         final String consumerKey = "xxx-xxx-xxx-xxx";
 
         //happy path - 204
@@ -296,9 +293,10 @@ public class WSO2ISKeyManagerImplTestCase {
     public void testRetrieveApplication() throws Exception {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
 
         //happy path - 200
         ////mocked response object from dcr api
@@ -375,9 +373,10 @@ public class WSO2ISKeyManagerImplTestCase {
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         OAuth2ServiceStubs.TokenServiceStub tokenStub = Mockito.mock(OAuth2ServiceStubs.TokenServiceStub.class);
         OAuth2ServiceStubs.RevokeServiceStub revokeStub = Mockito.mock(OAuth2ServiceStubs.RevokeServiceStub.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
 
         //happy path - 200 - password grant type
         ////request to key manager
@@ -428,9 +427,10 @@ public class WSO2ISKeyManagerImplTestCase {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         OAuth2ServiceStubs.TokenServiceStub tokenStub = Mockito.mock(OAuth2ServiceStubs.TokenServiceStub.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
 
         //happy path - 200 - refresh grant type
         ////request to key manager
@@ -468,9 +468,10 @@ public class WSO2ISKeyManagerImplTestCase {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         OAuth2ServiceStubs.TokenServiceStub tokenStub = Mockito.mock(OAuth2ServiceStubs.TokenServiceStub.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
 
         //happy path - 200 - client credentials grant type
         ////request to key manager
@@ -507,9 +508,10 @@ public class WSO2ISKeyManagerImplTestCase {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         OAuth2ServiceStubs.TokenServiceStub tokenStub = Mockito.mock(OAuth2ServiceStubs.TokenServiceStub.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
 
         //happy path - 200 - authorization code grant type
         ////request to key manager
@@ -549,9 +551,10 @@ public class WSO2ISKeyManagerImplTestCase {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         OAuth2ServiceStubs.TokenServiceStub tokenStub = Mockito.mock(OAuth2ServiceStubs.TokenServiceStub.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
 
         //happy path - 200 - JWT grant type
         ////request to key manager
@@ -591,9 +594,10 @@ public class WSO2ISKeyManagerImplTestCase {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         OAuth2ServiceStubs.TokenServiceStub tokenStub = Mockito.mock(OAuth2ServiceStubs.TokenServiceStub.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
 
         //error case - tokenRequest is null
         try {
@@ -667,9 +671,10 @@ public class WSO2ISKeyManagerImplTestCase {
         OAuth2ServiceStubs.IntrospectionServiceStub introspectionStub = Mockito.mock(
                 OAuth2ServiceStubs.IntrospectionServiceStub.class);
         Mockito.when(oAuth2ServiceStub.getIntrospectionServiceStub()).thenReturn(introspectionStub);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
         final String accessToken = "aaa-aaa-aaa-aaa";
 
         //happy path - 200 - token is active
@@ -753,14 +758,16 @@ public class WSO2ISKeyManagerImplTestCase {
         }
     }
 
-    @Test
+    //TODO:Enable after revoke endpoint implementation done in key manager.
+    @Test(enabled = false)
     public void testRevokeToken() throws Exception {
         DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
         OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
         OAuth2ServiceStubs.RevokeServiceStub revokeStub = Mockito.mock(OAuth2ServiceStubs.RevokeServiceStub.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
+        ScopeRegistration scopeRegistration = Mockito.mock
+                (ScopeRegistration.class);
+        DefaultKeyManagerImpl kmImpl = new DefaultKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
+                scopeRegistration);
 
         //happy path - 200
         Response revokeTokenResponse = Response.builder()
@@ -806,218 +813,6 @@ public class WSO2ISKeyManagerImplTestCase {
         }
     }
 
-    @Test
-    public void testAddScope() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        ScopeInfo scopeInfo = new ScopeInfo();
-        scopeInfo.setName("abc");
-        scopeInfo.setDescription("cde");
-        Mockito.when(scopeRegistrationServiceStub.registerScope(scopeInfo)).thenReturn(Response.builder().status(201)
-                .headers(new HashMap<>()).body(new Gson().toJson(scopeInfo), feign.Util.UTF_8)
-                .build());
-        Mockito.when(scopeRegistrationServiceStub.isScopeExist(scopeInfo.getName())).thenReturn(Response.builder()
-                .status(404).headers(Collections.emptyMap()).build());
-        Scope scope = new Scope();
-        scope.setName("abc");
-        scope.setDescription("cde");
-        Assert.assertTrue(kmImpl.registerScope(scope));
-    }
-
-    @Test
-    public void testAddScopeWhileScopeExist() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        ScopeInfo scopeInfo = new ScopeInfo();
-        scopeInfo.setName("abc");
-        scopeInfo.setDescription("cde");
-        Mockito.when(scopeRegistrationServiceStub.isScopeExist(scopeInfo.getName())).thenReturn(Response.builder()
-                .status(200).headers(Collections.emptyMap()).build());
-        Scope scope = new Scope();
-        scope.setName("abc");
-        scope.setDescription("cde");
-        Assert.assertFalse(kmImpl.registerScope(scope));
-    }
-
-    @Test
-    public void testScopeRegistrationFailed() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        ScopeInfo scopeInfo = new ScopeInfo();
-        scopeInfo.setName("abc");
-        scopeInfo.setDescription("cde");
-        Mockito.when(scopeRegistrationServiceStub.registerScope(scopeInfo)).thenReturn(Response.builder().status(400)
-                .headers(new HashMap<>()).body(new Gson().toJson(scopeInfo), feign.Util.UTF_8)
-                .build());
-        Mockito.when(scopeRegistrationServiceStub.isScopeExist(scopeInfo.getName())).thenReturn(Response.builder()
-                .status(404).headers(Collections.emptyMap()).build());
-        Scope scope = new Scope();
-        scope.setName("abc");
-        scope.setDescription("cde");
-        try {
-            kmImpl.registerScope(scope);
-            Assert.fail();
-        } catch (KeyManagementException e) {
-            Assert.assertEquals(e.getMessage(), "Scope Registration Failed");
-        }
-    }
-
-    @Test
-    public void testRetrievescope() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        ScopeInfo scopeInfo = new ScopeInfo();
-        scopeInfo.setName("abc");
-        scopeInfo.setDescription("cde");
-        Mockito.when(scopeRegistrationServiceStub.getScopeByName(scopeInfo.getName())).thenReturn(Response.builder()
-                .status(200).headers(new HashMap<>()).body(new Gson().toJson(scopeInfo), feign.Util.UTF_8).build());
-        Scope retrievedScope = kmImpl.retrieveScope(scopeInfo.getName());
-        Assert.assertEquals(scopeInfo.getName(), retrievedScope.getName());
-        Assert.assertEquals(scopeInfo.getDescription(), retrievedScope.getDescription());
-        Assert.assertEquals(Collections.emptyList(), retrievedScope.getBindings());
-    }
-
-    @Test
-    public void testRetrieveScopeWithBindings() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        ScopeInfo scopeInfo = new ScopeInfo();
-        scopeInfo.setName("abc");
-        scopeInfo.setDescription("cde");
-        List<String> list = new ArrayList<>();
-        list.add("apim:api_create");
-        scopeInfo.setBindings(list);
-        Mockito.when(scopeRegistrationServiceStub.getScopeByName(scopeInfo.getName())).thenReturn(Response.builder()
-                .status(200).headers(new HashMap<>()).body(new Gson().toJson(scopeInfo), feign.Util.UTF_8).build());
-        Scope retrievedScope = kmImpl.retrieveScope(scopeInfo.getName());
-        Assert.assertEquals(scopeInfo.getName(), retrievedScope.getName());
-        Assert.assertEquals(scopeInfo.getDescription(), retrievedScope.getDescription());
-        Assert.assertEquals(list, retrievedScope.getBindings());
-    }
-
-    @Test
-    public void testRetrieveNonExistingScope() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        Mockito.when(scopeRegistrationServiceStub.getScopeByName("abc")).thenReturn(Response.builder()
-                .status(404).headers(new HashMap<>()).build());
-        try {
-            kmImpl.retrieveScope("abc");
-            Assert.fail();
-        } catch (KeyManagementException e) {
-            Assert.assertEquals(e.getErrorHandler().getErrorCode(), 900981);
-
-        }
-    }
-
-    @Test
-    public void testRetrieveExistingScopeWhileBackendsendInternalServerError() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        Mockito.when(scopeRegistrationServiceStub.getScopeByName("abc")).thenReturn(Response.builder()
-                .status(500).headers(new HashMap<>()).build());
-        try {
-            kmImpl.retrieveScope("abc");
-            Assert.fail();
-        } catch (KeyManagementException e) {
-            Assert.assertEquals(e.getErrorHandler().getErrorCode(), 900967);
-
-        }
-    }
-
-    @Test
-    public void testUpdateScope() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        ScopeInfo scopeInfo = new ScopeInfo();
-        scopeInfo.setDescription("cde");
-        scopeInfo.setDisplayName("abc");
-        Scope scope = new Scope();
-        scope.setName("abc");
-        scope.setDescription("cde");
-        Mockito.when(scopeRegistrationServiceStub.updateScope(scopeInfo, scope.getName())).thenReturn(Response
-                .builder()
-                .status(200).headers(new HashMap<>()).body(new Gson().toJson(scopeInfo), feign.Util.UTF_8).build());
-        boolean status = kmImpl.updateScope(scope);
-        Assert.assertTrue(status);
-    }
-
-    @Test
-    public void testUpdateScopeWhileAuthorizationServerThrowsInternalError() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        ScopeInfo scopeInfo = new ScopeInfo();
-        scopeInfo.setDescription("cde");
-        scopeInfo.setDisplayName("abc");
-        Scope scope = new Scope();
-        scope.setName("abc");
-
-        scope.setDescription("cde");
-        Mockito.when(scopeRegistrationServiceStub.updateScope(scopeInfo, scope.getName())).thenReturn(Response
-                .builder()
-                .status(500).headers(new HashMap<>()).build());
-        try {
-            boolean status = kmImpl.updateScope(scope);
-        } catch (KeyManagementException e) {
-            Assert.assertEquals(e.getErrorHandler().getErrorCode(), 900967);
-        }
-    }
-    @Test
-    public void testDeleteScope() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        Mockito.when(scopeRegistrationServiceStub.deleteScope("abc")).thenReturn(Response
-                .builder()
-                .status(200).headers(new HashMap<>()).build());
-        boolean status = kmImpl.deleteScope("abc");
-        Assert.assertTrue(status);
-    }
-
-    @Test
-    public void testDeleteScopeWhileAuthorizationServerThrowsInternalError() throws APIManagementException {
-        DCRMServiceStub dcrmServiceStub = Mockito.mock(DCRMServiceStub.class);
-        OAuth2ServiceStubs oAuth2ServiceStub = Mockito.mock(OAuth2ServiceStubs.class);
-        ScopeRegistrationServiceStub scopeRegistrationServiceStub = Mockito.mock(ScopeRegistrationServiceStub.class);
-        WSO2ISKeyManagerImpl kmImpl = new WSO2ISKeyManagerImpl(dcrmServiceStub, oAuth2ServiceStub,
-                scopeRegistrationServiceStub);
-        Mockito.when(scopeRegistrationServiceStub.deleteScope("abc")).thenReturn(Response
-                .builder().status(500).headers(new HashMap<>()).build());
-        try {
-            boolean status = kmImpl.deleteScope("abc");
-        } catch (KeyManagementException e) {
-            Assert.assertEquals(e.getErrorHandler().getErrorCode(), 900983);
-        }
-    }
 
     private AccessTokenInfo createExpectedKeyManagerResponse(OAuth2TokenInfo oAuth2TokenInfo) {
         AccessTokenInfo accessTokenInfo = new AccessTokenInfo();
